+++ example_android/app/src/main/java/com/example/example_android/activity/SetFunctionDetailActivity.kt
@@ -46,9 +46,7 @@ class SetFunctionDetailActivity : BaseActivity() {
                     idoAlarmModel.name = "dddd"
                     idoAlarmModel.repeatTimes = 1;
                     idoBaseModel?.toJsonString()?.let { it1 -> Log.e("alarm11111", it1) }
-
                 }
-
             }
 
             CustomEvtType.SETWATCHDIALSORT -> {